refactor: layout에 있던 컴포넌트 page로 이동
