+++ frontend/techpick/src/components/folderContentHeader.css.ts
@@ -1,0 +1,11 @@
+import { style } from '@vanilla-extract/css';
+
+export const folderContentHeaderStyle = style({
+  display: 'flex',
+  flexDirection: 'column',
+  maxWidth: '1116px',
+  padding: '24px 36px',
+  width: '100%',
+  justifyContent: 'space-between',
+  transition: '0.3s ease',
+});
