+++ frontend/techpick/src/components/pickDraggableListLayout.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css';
+
+export const pickDraggableListLayoutStyle = style({
+  width: 'fit-content',
+  overflowY: 'scroll',
+});
