+++ frontend/techpick/src/components/qnAFloatingLink.css.ts
@@ -1,0 +1,8 @@
+import { style } from '@vanilla-extract/css';
+
+export const qnaSection = style({
+  position: 'absolute',
+  bottom: '32px',
+  right: '32px',
+});
+//
